ステージ開始処理を一本化

+++ Assets/Scripts/Stage/Stage.cs
@@ -143,16 +143,20 @@ public class Stage : Methods
     {
         visualizePlayer();
         sysPlayer.transform.localPosition = initialPlayerPosition;
-        sysPlayer.deleteArmorBar();
-        sysPlayer.setArmorBar();
-        sysPlayer.canRecieveKey = true;
 
-        Sys.playerHPbar = getBar(barType.HPbar, Color.red);
-        Sys.playerBRbar = getBar(barType.BRbar, Color.cyan);
-        Sys.playerENbar = getBar(barType.ENbar, Color.yellow);
-        Sys.playerHPbar.nowOrder = Order.publicState;
-        Sys.playerBRbar.nowOrder = Order.publicState;
-        Sys.playerENbar.nowOrder = Order.publicState;
+        if (!isSystem)
+        {
+            sysPlayer.deleteArmorBar();
+            sysPlayer.setArmorBar();
+            sysPlayer.canRecieveKey = true;
+
+            Sys.playerHPbar = getBar(barType.HPbar, Color.red);
+            Sys.playerBRbar = getBar(barType.BRbar, Color.cyan);
+            Sys.playerENbar = getBar(barType.ENbar, Color.yellow);
+            Sys.playerHPbar.nowOrder = Order.publicState;
+            Sys.playerBRbar.nowOrder = Order.publicState;
+            Sys.playerENbar.nowOrder = Order.publicState;
+        }
 
         StartCoroutine(nowStageAction = stageAction());
     }
